Merge from sam/manifests

+++ validate.sh
@@ -1,7 +1,16 @@
 #!/bin/bash
-echo "NOTE: If the docker run command gives you an error like 'server gave HTTP response to HTTPS client.' then you most likely are missing the insecure registry setting in Docker.  See https://confluence.internal.salesforce.com/x/NRDa (Set up Docker for Sam)"
+if [ "$1" == "evaluatePR" ] 
+then 
+    echo -e "\nEvaluating PR\n"
+    echo -e '```\n'
+    /opt/sam/sam-manifest-builder --root='./' -validateonly
+    echo -e '\n```\n'
+else
+  echo "NOTE: If the docker run command gives you an error like 'server gave HTTP response to HTTPS client.' then you most likely are missing the insecure registry setting in Docker.  See https://confluence.internal.salesforce.com/x/NRDa (Set up Docker for Sam)"
+  docker run -it --rm -v ${PWD}:/repo/ shared0-samcontrol1-1-prd.eng.sfdc.net:5000/hypersam:a9e68c1.clean.prahladjos-ltm.20161209_121757 /sam/sam-manifest-builder --root='/repo/' -validateonly
 
-docker run -it --rm -v ${PWD}:/repo/ shared0-samcontrol1-1-prd.eng.sfdc.net:5000/hypersam:a9e68c1.clean.prahladjos-ltm.20161209_121757 /sam/sam-manifest-builder --root='/repo/' -validateonly
+fi
+
 
 exitcode="$?"
 if [ "0" != "$exitcode" ]
